BackEnd/djact/requirements.txt
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
sqlparse==0.4.4
typing_extensions==4.10.0
pyotp==2.9.0
pypdf==4.0.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
tzlocal==5.2
uritools==4.0.2
urllib3==2.1.0
webencodings==0.5.1
xhtml2pdf==0.2.14
celery==5.4.0
redis==5.0.6
django-celery-beat== 2.6.0
django-celery-results==2.5.1
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
whitenoise==6.6.0
